tried making exe file
