+++ exe/bookroom.spec
@@ -7,7 +7,7 @@ block_cipher = None
 a = Analysis(['../proofOC/bookRoom.py'],
              pathex=[],
              binaries=[],
-             datas=[],
+             datas=[('../proofOC/RepoCommunicator.py', '.'), ('../proofOC/BookingBuilder.py', '.')],
              hiddenimports=[],
              hookspath=[],
              hooksconfig={},
