+++ proofOC/actions/main.yml
@@ -1,0 +1,33 @@
+
+# This is a basic workflow to help you get started with Actions
+
+name: Testing Automatin with Cronjob
+
+on:
+  schedule:
+  # Run “At 00:00.”
+    - cron: "0 0 * * *"
+    
+# A workflow run is made up of one or more jobs that can run sequentially or in parallel
+jobs:
+  build:
+    runs-on: macos-latest
+    
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+        - name: 🍽️ Checkout the repo
+          uses: actions/checkout@main
+          with:
+            fetch-depth: 1
+            
+        - name: 🐍 Set up Python 3.8
+          uses: actions/setup-python@v2
+          with:
+            python-version: '3.8'
+            
+        - name: Install Requirements
+          run: pip3 install -r scripts/requirements.txt
+
+        - name: Run Automation
+          run: python3 scripts/actionScript.py
+
