+++ proofOC/actions/requirements.txt
@@ -1,0 +1,2 @@
+selenium==4.1.3
+webdriver-manager==3.5.4
